[DLOG] Avoid double read in readahead
There was a bug in the ReadAheadReader whereby, if it was in a paused
state and the last segment it had read had been closed, then when a
writer started writing new segments, the ReadAheadReader would issue a
double read. This would end up triggering an error when it got to the
end of the segment as the issued one of the issued reads would then be
cancelled when the segment was closed.

This fix adds an explicit state machine to the reader and moves all
state transitions to run on the ordered executor, so only one read can
ever be issued at a time.

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #1973 from ivankelly/dlog-reader-crash-master
(cherry picked from commit db5d7e7cf30d58c9c463ce29c7d7900115717206)

Signed-off-by: Ivan Kelly <ivank@apache.org>