[DLOG] Sync zk before opening ledger
DLog uses a separate zookeeper session for dlog metadata and
bookkeeper metadata. This can introduce a race condition when a dlog
client creates a new segment and adds it to a log. If the dlog
metadata session has a newer view of the data than the bookkeeper
metadata session, then the dlog session will see there's a new segment
to be opened, but the zk session will throw a NoSuchLedger exception
when it actually tries to open it.

To solve this, we need to tell the bookkeeper zk session to ensure it
has the latest view of the data at the moment it tries to open a
ledger.